WIP: Tweaks and experiments

+++ src/player/Channel.h
@@ -96,7 +96,8 @@ class Channel {
                 }
                 _sampleIndex -= static_cast<float>(_sample->loopLength());
             }
-            *outputBuffer++ = (*_sample)[_sampleIndex] * _volume;
+            *outputBuffer++ =
+                (*_sample)[static_cast<size_t>(_sampleIndex)] * _volume;
             _sampleIndex += rate;
         }
     }
